added jni fallback

+++ de.ibapl.jnhw.posix/src/test/native/PollDefines.c
@@ -1,6 +1,6 @@
 /*
  * JNHW - Java Native header Wrapper, https://github.com/aploese/jnhw/
- * Copyright (C) 2019-2022, Arne Plöse and individual contributors as indicated
+ * Copyright (C) 2019-2023, Arne Plöse and individual contributors as indicated
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -24,7 +24,7 @@
 //We need the POSIX version ...
 #if !defined(HAVE_POLL_H) || !defined(_POSIX_VERSION)
 
-int getValueOf_HAVE_POLL_H() {
+int32_t getValueOf_HAVE_POLL_H() {
     return 0;
 }
 
@@ -32,47 +32,47 @@ int getValueOf_HAVE_POLL_H() {
 
 #include <poll.h>
 
-int getValueOf_HAVE_POLL_H() {
+int16_t getValueOf_HAVE_POLL_H() {
     return 1;
 }
 
-int getValueOf_POLLERR() {
+int16_t getValueOf_POLLERR() {
     return POLLERR;
 }
 
-int getValueOf_POLLHUP() {
+int16_t getValueOf_POLLHUP() {
     return POLLHUP;
 }
 
-int getValueOf_POLLIN() {
+int16_t getValueOf_POLLIN() {
     return POLLIN;
 }
 
-int getValueOf_POLLNVAL() {
+int16_t getValueOf_POLLNVAL() {
     return POLLNVAL;
 }
 
-int getValueOf_POLLOUT() {
+int16_t getValueOf_POLLOUT() {
     return POLLOUT;
 }
 
-int getValueOf_POLLPRI() {
+int16_t getValueOf_POLLPRI() {
     return POLLPRI;
 }
 
-int getValueOf_POLLRDBAND() {
+int16_t getValueOf_POLLRDBAND() {
     return POLLRDBAND;
 }
 
-int getValueOf_POLLRDNORM() {
+int16_t getValueOf_POLLRDNORM() {
     return POLLRDNORM;
 }
 
-int getValueOf_POLLWRBAND() {
+int16_t getValueOf_POLLWRBAND() {
     return POLLWRBAND;
 }
 
-int getValueOf_POLLWRNORM() {
+int16_t getValueOf_POLLWRNORM() {
     return POLLWRNORM;
 }
