+++ de.ibapl.jnhw.posix/src/test/native/SignalDefines.c
@@ -1,6 +1,6 @@
 /*
  * JNHW - Java Native header Wrapper, https://github.com/aploese/jnhw/
- * Copyright (C) 2019-2022, Arne Plöse and individual contributors as indicated
+ * Copyright (C) 2019-2023, Arne Plöse and individual contributors as indicated
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -24,134 +24,134 @@
     //We need the POSIX version ...
 #if !defined(HAVE_SIGNAL_H) || !defined(_POSIX_VERSION)
 
-int getValueOf_HAVE_SIGNAL_H() {
+int32_t getValueOf_HAVE_SIGNAL_H() {
     return 0;
 }
 #else
 #include <signal.h>
 #include <stdint.h>
 
-int getValueOf_HAVE_SIGNAL_H() {
+int32_t getValueOf_HAVE_SIGNAL_H() {
     return 1;
 }
 
-int getValueOf_ILL_ILLOPC() {
+int32_t getValueOf_ILL_ILLOPC() {
     return ILL_ILLOPC;
 }
 
-int getValueOf_ILL_ILLOPN() {
+int32_t getValueOf_ILL_ILLOPN() {
     return ILL_ILLOPN;
 }
 
-int getValueOf_ILL_ILLADR() {
+int32_t getValueOf_ILL_ILLADR() {
     return ILL_ILLADR;
 }
 
-int getValueOf_ILL_ILLTRP() {
+int32_t getValueOf_ILL_ILLTRP() {
     return ILL_ILLTRP;
 }
 
-int getValueOf_ILL_PRVOPC() {
+int32_t getValueOf_ILL_PRVOPC() {
     return ILL_PRVOPC;
 }
 
-int getValueOf_ILL_PRVREG() {
+int32_t getValueOf_ILL_PRVREG() {
     return ILL_PRVREG;
 }
 
-int getValueOf_ILL_COPROC() {
+int32_t getValueOf_ILL_COPROC() {
     return ILL_COPROC;
 }
 
-int getValueOf_ILL_BADSTK() {
+int32_t getValueOf_ILL_BADSTK() {
     return ILL_BADSTK;
 }
 
-int getValueOf_FPE_INTDIV() {
+int32_t getValueOf_FPE_INTDIV() {
     return FPE_INTDIV;
 }
 
-int getValueOf_FPE_INTOVF() {
+int32_t getValueOf_FPE_INTOVF() {
     return FPE_INTOVF;
 }
 
-int getValueOf_FPE_FLTDIV() {
+int32_t getValueOf_FPE_FLTDIV() {
     return FPE_FLTDIV;
 }
 
-int getValueOf_FPE_FLTOVF() {
+int32_t getValueOf_FPE_FLTOVF() {
     return FPE_FLTOVF;
 }
 
-int getValueOf_FPE_FLTUND() {
+int32_t getValueOf_FPE_FLTUND() {
     return FPE_FLTUND;
 }
 
-int getValueOf_FPE_FLTRES() {
+int32_t getValueOf_FPE_FLTRES() {
     return FPE_FLTRES;
 }
 
-int getValueOf_FPE_FLTINV() {
+int32_t getValueOf_FPE_FLTINV() {
     return FPE_FLTINV;
 }
 
-int getValueOf_FPE_FLTSUB() {
+int32_t getValueOf_FPE_FLTSUB() {
     return FPE_FLTSUB;
 }
 
-int getValueOf_SEGV_MAPERR() {
+int32_t getValueOf_SEGV_MAPERR() {
     return SEGV_MAPERR;
 }
 
-int getValueOf_SEGV_ACCERR() {
+int32_t getValueOf_SEGV_ACCERR() {
     return SEGV_ACCERR;
 }
 
-int getValueOf_BUS_ADRALN() {
+int32_t getValueOf_BUS_ADRALN() {
     return BUS_ADRALN;
 }
 
-int getValueOf_BUS_ADRERR() {
+int32_t getValueOf_BUS_ADRERR() {
     return BUS_ADRERR;
 }
 
-int getValueOf_BUS_OBJERR() {
+int32_t getValueOf_BUS_OBJERR() {
     return BUS_OBJERR;
 }
 
-int getValueOf_TRAP_BRKPT() {
+int32_t getValueOf_TRAP_BRKPT() {
     return TRAP_BRKPT;
 }
 
-int getValueOf_TRAP_TRACE() {
+int32_t getValueOf_TRAP_TRACE() {
     return TRAP_TRACE;
 }
 
-int getValueOf_CLD_EXITED() {
+int32_t getValueOf_CLD_EXITED() {
     return CLD_EXITED;
 }
 
-int getValueOf_CLD_KILLED() {
+int32_t getValueOf_CLD_KILLED() {
     return CLD_KILLED;
 }
 
-int getValueOf_CLD_DUMPED() {
+int32_t getValueOf_CLD_DUMPED() {
     return CLD_DUMPED;
 }
 
-int getValueOf_CLD_TRAPPED() {
+int32_t getValueOf_CLD_TRAPPED() {
     return CLD_TRAPPED;
 }
 
-int getValueOf_CLD_STOPPED() {
+int32_t getValueOf_CLD_STOPPED() {
     return CLD_STOPPED;
 }
 
-int getValueOf_CLD_CONTINUED() {
+int32_t getValueOf_CLD_CONTINUED() {
     return CLD_CONTINUED;
 }
 
-int* tryGetValueOf_POLL_IN(int* value) {
+int32_t* tryGetValueOf_POLL_IN(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_IN)
     value = NULL;
@@ -164,7 +164,7 @@ int* tryGetValueOf_POLL_IN(int* value) {
     return value;
 }
 
-int* tryGetValueOf_POLL_OUT(int* value) {
+int32_t* tryGetValueOf_POLL_OUT(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_OUT)
     value = NULL;
@@ -177,7 +177,7 @@ int* tryGetValueOf_POLL_OUT(int* value) {
     return value;
 }
 
-int* tryGetValueOf_POLL_MSG(int* value) {
+int32_t* tryGetValueOf_POLL_MSG(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_MSG)
     value = NULL;
@@ -190,7 +190,7 @@ int* tryGetValueOf_POLL_MSG(int* value) {
     return value;
 }
 
-int* tryGetValueOf_POLL_ERR(int* value) {
+int32_t* tryGetValueOf_POLL_ERR(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_ERR)
     value = NULL;
@@ -203,7 +203,7 @@ int* tryGetValueOf_POLL_ERR(int* value) {
     return value;
 }
 
-int* tryGetValueOf_POLL_PRI(int* value) {
+int32_t* tryGetValueOf_POLL_PRI(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_PRI)
     value = NULL;
@@ -216,7 +216,7 @@ int* tryGetValueOf_POLL_PRI(int* value) {
     return value;
 }
 
-int* tryGetValueOf_POLL_HUP(int* value) {
+int32_t* tryGetValueOf_POLL_HUP(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(POLL_HUP)
     value = NULL;
@@ -229,19 +229,19 @@ int* tryGetValueOf_POLL_HUP(int* value) {
     return value;
 }
 
-int getValueOf_SI_USER() {
+int32_t getValueOf_SI_USER() {
     return SI_USER;
 }
 
-int getValueOf_SI_QUEUE() {
+int32_t getValueOf_SI_QUEUE() {
     return SI_QUEUE;
 }
 
-int getValueOf_SI_TIMER() {
+int32_t getValueOf_SI_TIMER() {
     return SI_TIMER;
 }
 
-int* tryGetValueOf_SI_ASYNCIO(int* value) {
+int32_t* tryGetValueOf_SI_ASYNCIO(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SI_ASYNCIO)
     value = NULL;
@@ -254,7 +254,7 @@ int* tryGetValueOf_SI_ASYNCIO(int* value) {
     return value;
 }
 
-int* tryGetValueOf_SI_MESGQ(int* value) {
+int32_t* tryGetValueOf_SI_MESGQ(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SI_MESGQ)
     value = NULL;
@@ -267,73 +267,73 @@ int* tryGetValueOf_SI_MESGQ(int* value) {
     return value;
 }
 
-int getValueOf_SIG_BLOCK() {
+int32_t getValueOf_SIG_BLOCK() {
     return SIG_BLOCK;
 }
 
-int getValueOf_SIG_UNBLOCK() {
+int32_t getValueOf_SIG_UNBLOCK() {
     return SIG_UNBLOCK;
 }
 
-int getValueOf_SIG_SETMASK() {
+int32_t getValueOf_SIG_SETMASK() {
     return SIG_SETMASK;
 }
 
-int getValueOf_SA_NOCLDSTOP() {
+int32_t getValueOf_SA_NOCLDSTOP() {
     return SA_NOCLDSTOP;
 }
 
-int getValueOf_SA_ONSTACK() {
+int32_t getValueOf_SA_ONSTACK() {
     return SA_ONSTACK;
 }
 
-int getValueOf_SA_RESETHAND() {
+int32_t getValueOf_SA_RESETHAND() {
     return (int32_t) SA_RESETHAND;
 }
 
-int getValueOf_SA_RESTART() {
+int32_t getValueOf_SA_RESTART() {
     return SA_RESTART;
 }
 
-int getValueOf_SA_SIGINFO() {
+int32_t getValueOf_SA_SIGINFO() {
     return SA_SIGINFO;
 }
 
-int getValueOf_SA_NOCLDWAIT() {
+int32_t getValueOf_SA_NOCLDWAIT() {
     return SA_NOCLDWAIT;
 }
 
-int getValueOf_SA_NODEFER() {
+int32_t getValueOf_SA_NODEFER() {
     return SA_NODEFER;
 }
 
-int getValueOf_SS_ONSTACK() {
+int32_t getValueOf_SS_ONSTACK() {
     return SS_ONSTACK;
 }
 
-int getValueOf_SS_DISABLE() {
+int32_t getValueOf_SS_DISABLE() {
     return SS_DISABLE;
 }
 
 //glibc > 2.34 redefines this to sysconf(_SC_MINSIGSTKSZ), so this const is useless
-long getValueOf_MINSIGSTKSZ() {
+int64_t getValueOf_MINSIGSTKSZ() {
     return MINSIGSTKSZ;
 }
 
 //glibc > 2.34 redefines this to sysconf(_SC_SIGSTKSZ), so this const is useless
-long getValueOf_SIGSTKSZ() {
+int64_t getValueOf_SIGSTKSZ() {
     return SIGSTKSZ;
 }
 
-long getValueOf_SIG_DFL() {
+intptr_t getValueOf_SIG_DFL() {
     return (intptr_t) SIG_DFL;
 }
 
-long getValueOf_SIG_ERR() {
+intptr_t getValueOf_SIG_ERR() {
     return (intptr_t) SIG_ERR;
 }
 
-int* tryGetValueOf_SIG_HOLD(int* value) {
+intptr_t* tryGetValueOf_SIG_HOLD(intptr_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SIG_HOLD)
      value = NULL;
@@ -346,11 +346,11 @@ int* tryGetValueOf_SIG_HOLD(int* value) {
     return value;
 }
 
-long getValueOf_SIG_IGN() {
+intptr_t getValueOf_SIG_IGN() {
     return (intptr_t) SIG_IGN;
 }
 
-int* tryGetValueOf_SIGEV_NONE(int* value) {
+int32_t* tryGetValueOf_SIGEV_NONE(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SIGEV_NONE)
     value = NULL;
@@ -363,7 +363,7 @@ int* tryGetValueOf_SIGEV_NONE(int* value) {
     return value;
 }
 
-int* tryGetValueOf_SIGEV_SIGNAL(int* value) {
+int32_t* tryGetValueOf_SIGEV_SIGNAL(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SIGEV_SIGNAL)
     value = NULL;
@@ -376,7 +376,7 @@ int* tryGetValueOf_SIGEV_SIGNAL(int* value) {
     return value;
 }
 
-int* tryGetValueOf_SIGEV_THREAD(int* value) {
+int32_t* tryGetValueOf_SIGEV_THREAD(int32_t* value) {
 #if defined(__OpenBSD__)
 #if !defined(SIGEV_THREAD)
     value = NULL;
@@ -389,87 +389,87 @@ int* tryGetValueOf_SIGEV_THREAD(int* value) {
     return value;
 }
 
-int getValueOf_SIGABRT() {
+int32_t getValueOf_SIGABRT() {
     return SIGABRT;
 }
 
-int getValueOf_SIGALRM() {
+int32_t getValueOf_SIGALRM() {
     return SIGALRM;
 }
 
-int getValueOf_SIGBUS() {
+int32_t getValueOf_SIGBUS() {
     return SIGBUS;
 }
 
-int getValueOf_SIGCHLD() {
+int32_t getValueOf_SIGCHLD() {
     return SIGCHLD;
 }
 
-int getValueOf_SIGCONT() {
+int32_t getValueOf_SIGCONT() {
     return SIGCONT;
 }
 
-int getValueOf_SIGFPE() {
+int32_t getValueOf_SIGFPE() {
     return SIGFPE;
 }
 
-int getValueOf_SIGHUP() {
+int32_t getValueOf_SIGHUP() {
     return SIGHUP;
 }
 
-int getValueOf_SIGILL() {
+int32_t getValueOf_SIGILL() {
     return SIGILL;
 }
 
-int getValueOf_SIGINT() {
+int32_t getValueOf_SIGINT() {
     return SIGINT;
 }
 
-int getValueOf_SIGKILL() {
+int32_t getValueOf_SIGKILL() {
     return SIGKILL;
 }
 
-int getValueOf_SIGPIPE() {
+int32_t getValueOf_SIGPIPE() {
     return SIGPIPE;
 }
 
-int getValueOf_SIGQUIT() {
+int32_t getValueOf_SIGQUIT() {
     return SIGQUIT;
 }
 
-int getValueOf_SIGSEGV() {
+int32_t getValueOf_SIGSEGV() {
     return SIGSEGV;
 }
 
-int getValueOf_SIGSTOP() {
+int32_t getValueOf_SIGSTOP() {
     return SIGSTOP;
 }
 
-int getValueOf_SIGTERM() {
+int32_t getValueOf_SIGTERM() {
     return SIGTERM;
 }
 
-int getValueOf_SIGTSTP() {
+int32_t getValueOf_SIGTSTP() {
     return SIGTSTP;
 }
 
-int getValueOf_SIGTTIN() {
+int32_t getValueOf_SIGTTIN() {
     return SIGTTIN;
 }
 
-int getValueOf_SIGTTOU() {
+int32_t getValueOf_SIGTTOU() {
     return SIGTTOU;
 }
 
-int getValueOf_SIGUSR1() {
+int32_t getValueOf_SIGUSR1() {
     return SIGUSR1;
 }
 
-int getValueOf_SIGUSR2() {
+int32_t getValueOf_SIGUSR2() {
     return SIGUSR2;
 }
 
-int* tryGetValueOf_SIGPOLL(int* value) {
+int32_t* tryGetValueOf_SIGPOLL(int32_t* value) {
 #if defined(__APPLE__) || defined(__FreeBSD__) || defined(__OpenBSD__)
 #if !defined(SIGPOLL)
     value = NULL;
@@ -482,31 +482,31 @@ int* tryGetValueOf_SIGPOLL(int* value) {
     return value;
 }
 
-int getValueOf_SIGPROF() {
+int32_t getValueOf_SIGPROF() {
     return SIGPROF;
 }
 
-int getValueOf_SIGSYS() {
+int32_t getValueOf_SIGSYS() {
     return SIGSYS;
 }
 
-int getValueOf_SIGTRAP() {
+int32_t getValueOf_SIGTRAP() {
     return SIGTRAP;
 }
 
-int getValueOf_SIGURG() {
+int32_t getValueOf_SIGURG() {
     return SIGURG;
 }
 
-int getValueOf_SIGVTALRM() {
+int32_t getValueOf_SIGVTALRM() {
     return SIGVTALRM;
 }
 
-int getValueOf_SIGXCPU() {
+int32_t getValueOf_SIGXCPU() {
     return SIGXCPU;
 }
 
-int getValueOf_SIGXFSZ() {
+int32_t getValueOf_SIGXFSZ() {
     return SIGXFSZ;
 }
